fix: add default compiler to server dir
