+++ packages/server/plugin-server/package.json
@@ -57,7 +57,6 @@
     }
   },
   "dependencies": {
-    "@modern-js/server-utils": "workspace:*",
     "@modern-js/utils": "workspace:*",
     "@swc/helpers": "0.5.3"
   },
