+++ packages/solutions/app-tools/package.json
@@ -90,6 +90,7 @@
     "@modern-js/plugin-lint": "workspace:*",
     "@modern-js/prod-server": "workspace:*",
     "@modern-js/server": "workspace:*",
+    "@modern-js/server-utils": "workspace:*",
     "@modern-js/types": "workspace:*",
     "@modern-js/upgrade": "workspace:*",
     "@modern-js/utils": "workspace:*",
